fix chunker issue

+++ internal/ingestion/component/chunker/title_test.go
@@ -446,8 +446,8 @@ func TestOutlineFromInputs(t *testing.T) {
 	inputs := map[string]any{
 		"file": map[string]any{
 			"outline": []any{
-				map[string]any{"title": "第一章", "level": 0, "page_number": 1},
-				map[string]any{"title": "第二章", "level": 1, "page_number": 3},
+				map[string]any{"title": "第一章", "level": float64(0), "page_number": float64(1)},
+				map[string]any{"title": "第二章", "level": float64(1), "page_number": float64(3)},
 			},
 		},
 	}
